requirements.txt
asn1crypto==0.24.0
backports.csv==1.0.7
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.2
django-cors-headers==3.0.2
django-crispy-forms==1.7.2
django-filter==2.1.0
django-formtools==2.1
django-guardian==2.0.0
django-import-export==1.2.0
django-qiniu-storage==2.3.1
django-redis==4.10.0
django-reversion==3.0.4
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
drf-extensions==0.5.0
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.13.0
idna==2.7
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
Markdown==3.1.1
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
Naked==0.1.31
oauthlib==3.0.1
odfpy==1.4.0
openpyxl==2.6.2
Pillow==6.0.0
pycparser==2.19
pycryptodome==3.8.2
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.1
PyYAML==5.1.1
qiniu==7.2.4
raven==6.10.0
redis==3.2.1
requests==2.22.0
requests-oauthlib==1.2.0
shellescape==3.4.1
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
tablib==0.13.0
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.25.3
xlrd==1.2.0
xlwt==1.3.0
